Release 1.0

+++ src/components/links/FolderMenuList/FolderMenu.tsx
@@ -3,6 +3,7 @@ import Image from 'next/image';
 import { CommonButton, CommonModal, ModalRenderer } from '@/components';
 import { FolderContext, useModal } from '@/lib/context';
 import { deleteFolder, getFolderList, updateFolder } from '@/lib/api';
+import { useLinksContextSelector } from '@/components/links-component';
 
 type ModalType = 'add' | 'share' | 'changeName' | 'delete';
 
@@ -15,15 +16,50 @@ interface FolderMenuProps {
 export const FolderMenu = ({ src, text, modalType }: FolderMenuProps) => {
   const { setFolderList, selectedFolder, setSelectedFolder } =
     useContext(FolderContext);
+  const {
+    linksAction: { data: linkData },
+  } = useLinksContextSelector();
   const { openModal } = useModal();
-
   const [newFolderName, setNewFolderName] = useState('');
   const getInputValue = (e: React.ChangeEvent<HTMLInputElement>) => {
     setNewFolderName(e.target.value);
   };
 
   const [showModal, setShowModal] = useState<boolean>(false);
-  const handleCloseModal = () => setShowModal((prev) => !prev);
+  const handleCloseModal = () => {
+    let modalTypeText = '';
+    switch (modalType) {
+      case 'share':
+        modalTypeText = '공유';
+        break;
+      case 'changeName':
+        modalTypeText = '변경';
+        break;
+      case 'delete':
+        modalTypeText = '삭제';
+        break;
+      default:
+        modalTypeText = '';
+        break;
+    }
+    if (selectedFolder.id === 0) {
+      openModal({
+        type: 'alert',
+        key: `preventDefaultFolder${modalType}`,
+        message: `전체 폴더는 ${modalTypeText}할 수 없습니다.`,
+      });
+      return;
+    }
+    if (modalType === 'share' && linkData && linkData.data.totalCount < 1) {
+      openModal({
+        type: 'alert',
+        key: 'preventEmptyFolderShare',
+        message: '폴더가 비어있습니다. 먼저 링크를 추가해주세요 😄',
+      });
+      return;
+    }
+    setShowModal((prev) => !prev);
+  };
 
   const fetchFolderList = async () => {
     const data = await getFolderList();
@@ -48,6 +84,17 @@ export const FolderMenu = ({ src, text, modalType }: FolderMenuProps) => {
 
   const handleDeleteButtonClick = async () => {
     const folderId = selectedFolder.id;
+    if (linkData && linkData.data.totalCount > 0) {
+      openModal({
+        type: 'alert',
+        key: 'notEmptyFolderDeleteError',
+        message: '링크가 들어있는 폴더는 삭제할 수 없습니다.',
+        onConfirm() {
+          setShowModal(false);
+        },
+      });
+      return;
+    }
     try {
       await deleteFolder({ folderId });
       fetchFolderList();
